Modules,Comments &pip
1. Modules
A module is a file containing Python code (functions, classes, variables).
It allows you to reuse code by importing it into other scripts.

2. Comments
Comments are notes in the code that explain what the code does. They are ignored by Python.

3. pip
pip is a tool to install and manage Python packages (like requests, pyjokes).